refactor: Resource (pack) loading parts are now in a separate file
